voting_page/requirements.txt
pillow==10.3.0
gspread
pandas